+++ R/pkgdown.R
@@ -4,9 +4,8 @@
 #' [pkgdown](https://pkgdown.r-lib.org) makes it easy to turn your package into
 #' a beautiful website. usethis provides two functions to help you use pkgdown:
 #'
-#' * `use_pkgdown()`: creates a pkgdown config file, adds relevant files or
-#'   directories to `.Rbuildignore` and `.gitignore`, and builds favicons if
-#'   your package has a logo.
+#' * `use_pkgdown()`: creates a pkgdown config file and adds relevant files or
+#'   directories to `.Rbuildignore` and `.gitignore`.
 #'
 #' * `use_pkgdown_github_pages()`: implements the GitHub setup needed to
 #'   automatically publish your pkgdown site to GitHub pages:
@@ -31,28 +30,37 @@ use_pkgdown <- function(config_file = "_pkgdown.yml", destdir = "docs") {
   check_is_package("use_pkgdown()")
   check_installed("pkgdown")
 
-  use_build_ignore(c(config_file, destdir))
-  use_build_ignore("pkgdown")
+  use_build_ignore(c(config_file, destdir, "pkgdown"))
   use_git_ignore(destdir)
 
-  ui_todo("
-    Record your site's {ui_field('url')} in the pkgdown config file \\
-    (optional, but recommended)")
-
-  if (has_logo()) {
-    pkgdown_build_favicons(proj_get(), overwrite = TRUE)
-  }
-
-  config <- proj_path(config_file)
-  if (!identical(destdir, "docs")) {
-    write_over(config, paste("destination:", destdir))
-  }
-  edit_file(config)
+  config <- pkgdown_config(destdir)
+  config_path <- proj_path(config_file)
+  write_over(config_path, yaml::as.yaml(config))
+  edit_file(config_path)
 
   invisible(TRUE)
 }
 
-# tidyverse pkgdown setup ------------------------------------------------------
+pkgdown_config <- function(destdir) {
+  config <- list(
+    url = NULL
+  )
+
+  if (pkgdown_version() >= "1.9000") {
+    config$template <- list(bootstrap = 4L)
+  }
+
+  if (!identical(destdir, "docs")) {
+    config$destination <- destdir
+  }
+
+  config
+}
+
+# wrapping because I need to be able to mock this in tests
+pkgdown_version <- function() {
+  utils::packageVersion("pkgdown")
+}
 
 #' @rdname use_pkgdown
 #' @export
@@ -75,6 +83,8 @@ use_pkgdown_github_pages <- function() {
   }
 }
 
+# tidyverse pkgdown setup ------------------------------------------------------
+
 #' @details
 #' * `use_tidy_pkgdown_github_pages()` is basically
 #'   [use_pkgdown_github_pages()], so does full pkgdown set up. Note that there
@@ -94,32 +104,22 @@ use_tidy_pkgdown_github_pages <- function() {
 use_pkgdown_url <- function(url, tr = NULL) {
   tr <- tr %||% target_repo(github_get = TRUE)
 
-  config <- pkgdown_config_path()
-  config_lines <- read_utf8(config)
-  url_line <- paste0("url: ", url)
-  if (!any(grepl(url_line, config_lines))) {
-    ui_done("
-      Recording {ui_value(url)} as site's {ui_field('url')} in \\
-      {ui_path(config)}")
-    config_lines <- config_lines[!grepl("^url:", config_lines)]
-    write_utf8(config, c(
-      url_line,
-      if (length(config_lines) && nzchar(config_lines[[1]])) "",
-      config_lines
-    ))
+  config_path <- pkgdown_config_path()
+  ui_done("
+    Recording {ui_value(url)} as site's {ui_field('url')} in \\
+    {ui_path(config_path)}")
+  config <- pkgdown_config_meta()
+  if (has_name(config, "url")) {
+    config$url <- url
+  } else {
+    config <- c(url = url, config)
   }
+  write_utf8(config_path, yaml::as.yaml(config))
 
-  urls <- desc::desc_get_urls()
-  if (!url %in% urls) {
-    ui_done("Adding {ui_value(url)} to {ui_field('URL')} field in DESCRIPTION")
-    ui_silence(
-      use_description_field(
-        "URL",
-        glue_collapse(c(url, urls), ", "),
-        overwrite = TRUE
-      )
-    )
-  }
+  ui_done("Adding {ui_value(url)} to {ui_field('URL')} field in DESCRIPTION")
+  desc <- desc::desc(file = proj_get())
+  desc$add_urls(url)
+  desc$write()
 
   gh <- gh_tr(tr)
   homepage <- gh("GET /repos/{owner}/{repo}")[["homepage"]]
@@ -215,15 +215,10 @@ use_pkgdown_travis <- function() {
 
   tr <- target_repo(github_get = TRUE)
 
-  use_build_ignore("docs/")
+  use_build_ignore(c("docs/", "pkgdown"))
   use_git_ignore("docs/")
   # TODO: suggest git rm -r --cache docs/
   # Can't currently detect if git known files in that directory
-
-  if (has_logo()) {
-    pkgdown_build_favicons(proj_get(), overwrite = TRUE)
-    use_build_ignore("pkgdown")
-  }
 
   ui_todo("
     Set up deploy keys by running {ui_code('travis::use_travis_deploy()')}")
@@ -241,10 +236,4 @@ use_pkgdown_travis <- function() {
   use_github_pages()
 
   invisible()
-}
-
-# usethis itself should not depend on pkgdown
-# all usage of this wrapper is guarded by `check_installed("pkgdown")`
-pkgdown_build_favicons <- function(...) {
-  get("build_favicons", asNamespace("pkgdown"), mode = "function")(...)
 }